+++ lib/tests/unit/messages/ReadHoldingRegistersResponseImplTest.cpp
@@ -92,7 +92,7 @@ TEST_CASE("ReadHoldingRegistersResponseImpl")
                 0x00, 0x03, // Register 0x0018
                 0x00, 0x04  // Register 0x0019
             }};
-            ser4cpp::rseq_t buffer{proper_response.data(), proper_response.size()};
+            ser4cpp::rseq_t buffer{proper_response.data(), static_cast<uint32_t>(proper_response.size())};
 
             auto result = ReadHoldingRegistersResponseImpl::parse(request_impl, buffer);
 
@@ -116,7 +116,7 @@ TEST_CASE("ReadHoldingRegistersResponseImpl")
                 0x83, // Exception function code
                 0x02  // Illegal data address
             }};
-            ser4cpp::rseq_t buffer{exception_response.data(), exception_response.size()};
+            ser4cpp::rseq_t buffer{exception_response.data(), static_cast<uint32_t>(exception_response.size())};
 
             auto result = ReadHoldingRegistersResponseImpl::parse(request_impl, buffer);
 
@@ -129,7 +129,7 @@ TEST_CASE("ReadHoldingRegistersResponseImpl")
             std::array<uint8_t, 1> wrong_size_response{ {
                 0x03 // Function code
             }};
-            ser4cpp::rseq_t buffer{ wrong_size_response.data(), wrong_size_response.size() };
+            ser4cpp::rseq_t buffer{wrong_size_response.data(), static_cast<uint32_t>(wrong_size_response.size())};
 
             auto result = ReadHoldingRegistersResponseImpl::parse(request_impl, buffer);
 
@@ -143,7 +143,7 @@ TEST_CASE("ReadHoldingRegistersResponseImpl")
                 0x03,            // Size not even
                 0x42, 0x43, 0x44 // Appropriate data
             }};
-            ser4cpp::rseq_t buffer{ wrong_size_response.data(), wrong_size_response.size() };
+            ser4cpp::rseq_t buffer{wrong_size_response.data(), static_cast<uint32_t>(wrong_size_response.size())};
 
             auto result = ReadHoldingRegistersResponseImpl::parse(request_impl, buffer);
 
@@ -157,7 +157,7 @@ TEST_CASE("ReadHoldingRegistersResponseImpl")
                 0x02, // Length should be 2
                 0x42  // Only 1 byte is received
             }};
-            ser4cpp::rseq_t buffer{wrong_size_response.data(), wrong_size_response.size()};
+            ser4cpp::rseq_t buffer{wrong_size_response.data(), static_cast<uint32_t>(wrong_size_response.size())};
 
             auto result = ReadHoldingRegistersResponseImpl::parse(request_impl, buffer);
